AIService/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
decorator==5.1.1
ecdsa==0.18.0
fastapi==0.95.1
fastapi-security==0.5.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
joblib==1.2.0
multidict==6.0.4
numpy==1.24.3
openai==0.27.6
Pillow==9.5.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
requests==2.28.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
validators==0.20.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
